Banking app - added delete operations, RequestPOJO and updated logics

+++ banking-app/src/main/java/com/example/bankingapp/service/BankingAppService.java
@@ -1,16 +1,18 @@
 package com.example.bankingapp.service;
 
+import com.example.bankingapp.exceptions.NoUserExistsException;
+import com.example.bankingapp.exceptions.UserExistsException;
+import com.example.bankingapp.exceptions.UserNotFoundException;
+import com.example.bankingapp.utils.RequestPOJO;
 import com.example.bankingapp.dao.AccountDetailsRepository;
 import com.example.bankingapp.dao.UserRepository;
 import com.example.bankingapp.entity.AccountDetails;
-import com.example.bankingapp.entity.AccountType;
 import com.example.bankingapp.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class BankingAppService {
@@ -23,28 +25,55 @@ public class BankingAppService {
         this.accountDetailsRepository = accountDetailsRepository;
     }
 
-    public AccountDetails createOrUpdate(User user) {
-        userRepository.save(user);
-        AccountDetails accountDetails = new AccountDetails();
+    public AccountDetails createOrUpdate(RequestPOJO requestPOJO) {
+        if (userRepository.existsByEmail(requestPOJO.getEmail())) {
+            throw new UserExistsException("User already exists with the email ID");
+        }
+        AccountDetails accountDetails = requestPOJO.toAccountDetails();
+        User user = userRepository.save(accountDetails.getUser());
         accountDetails.setUser(user);
-        accountDetails.setAccountType(user.getAccountType());
-        accountDetails.setBalance(user.getAccountType()== AccountType.CURRENT ? 500.0 : 0.0);
-        accountDetails = accountDetailsRepository.save(accountDetails);
+        accountDetailsRepository.save(accountDetails);
         return accountDetails;
     }
 
-    public User findUser(long userId) {
-        Optional<User> user = userRepository.findById(userId);
-        return user.orElse(null);
-    }
-
-    public List<AccountDetails> deleteAll() {
-        List<AccountDetails> accountDetails = accountDetailsRepository.findAll();
-        accountDetailsRepository.deleteAll();
+    public AccountDetails findUser(long userId) {
+        Optional<User> optionalUser = userRepository.findById(userId);
+        User user = optionalUser.orElse(null);
+        if (user == null) {
+            throw new UserNotFoundException("No user present with given user id") ;
+        }
+        AccountDetails accountDetails = accountDetailsRepository.findByUserId(userId);
+        accountDetails.setUser(user);
         return accountDetails;
     }
 
     public List<AccountDetails> findAll() {
-        return accountDetailsRepository.findAll();
+        List<AccountDetails> accountDetails = accountDetailsRepository.findAll();
+        if (accountDetails.isEmpty()) {
+            throw new NoUserExistsException("No user present in the database");
+        }
+        return accountDetails;
+    }
+
+    public List<AccountDetails> deleteAll() {
+        List<AccountDetails> accountDetails = accountDetailsRepository.findAll();
+        if (accountDetails.isEmpty()) {
+            throw new NoUserExistsException("No user present in the database");
+        }
+        accountDetailsRepository.deleteAll();
+        userRepository.deleteAll();
+        return accountDetails;
+    }
+
+    public AccountDetails deleteById(long userId) {
+        User user = userRepository.findById(userId).orElse(null);
+        if (user == null) {
+            throw new UserNotFoundException("No user present with given user id") ;
+        }
+        AccountDetails accountDetails = accountDetailsRepository.findByUserId(userId);
+        accountDetails.setUser(user);
+        accountDetailsRepository.deleteByUserId(userId);
+        userRepository.deleteById(userId);
+        return accountDetails;
     }
 }
